Combined flash and app engine elements of the project
Moved the app engine files into the root directory and moved the flash files
into a subdirectory.  Updated the app engine chat.html file to pull in the flash
content via an 'include' template tag.  This allows the Flex developer(s) to
continue working on and updating the SWF independently, while changes they make
will be pulled into the chat.html file without any further copy/pasting.

Also rewrote the SWF to get rid of the URL-based flow of connecting to another
user, and added in a form-based flow where variables are passed to the SWF
via Javascript.  Made a few other changes to the SWF to better align with the
new flow.