Removed old files from working tree and updated the index.html form
Hadn't actually removed the files in my last commit, so did that this time.
Also updated the form in index.html to make languages a select field instead
of a checkbox, and to remove the option to hide yourself from the list.
Made the corresponding changes in the main.py file as well.